move /src/main.py to /main.py, change other imports accordingly, remove root.py contents to info.py
