+++ src/managers/devicemanager.py
@@ -1,5 +1,5 @@
 import platform
-from managers.pciids import PCIIDs
+from src.managers.pciids import PCIIDs
 
 
 class DeviceManager:
@@ -21,15 +21,15 @@ class DeviceManager:
         self.logger = logger
 
         if self.platform == "darwin":
-            from dumps.macOS.mac import MacHardwareManager
+            from src.dumps.macOS.mac import MacHardwareManager
 
             self.manager = MacHardwareManager(self)
         elif self.platform == "linux":
-            from dumps.Linux.linux import LinuxHardwareManager
+            from src.dumps.Linux.linux import LinuxHardwareManager
 
             self.manager = LinuxHardwareManager(self)
         elif self.platform == "windows":
-            from dumps.Windows.win import WindowsHardwareManager
+            from src.dumps.Windows.win import WindowsHardwareManager
 
             self.manager = WindowsHardwareManager(self)
