--- src/root.py
@@ -1,3 +1,0 @@
-import os
-
-root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
