+++ src/util/missing_dep.py
@@ -1,13 +1,12 @@
+import os.path
 from unittest import TestCase
 from pkg_resources import require, DistributionNotFound
 from subprocess import call
 from sys import platform, executable
-from pathlib import Path
+from src import info
 
 try:
-    REQUIRED = (
-        Path(Path(__file__).parent).parent.with_name("requirements.txt").open().read()
-    )
+    REQUIRED = open(os.path.join(info.root_dir, "requirements.txt")).read()
 except Exception as e:
     raise Exception(
         f"Failed to locate requirements file. Maybe it was deleted?\n\n{str(e)}"
